add CMake, remove qmake. change QtGifImage

+++ src/GIFCreation/gifLib/dgif_lib.c
@@ -28,9 +28,9 @@ two modules will be linked.  Preserve this property!
 #define UNSIGNED_LITTLE_ENDIAN(lo, hi) ((lo) | ((hi) << 8))
 
 /* avoid extra function call in case we use fread (TVT) */
-#define READ(_gif, _buf, _len)                                                 \
-  (((GifFilePrivateType *)_gif->Private)->Read                                 \
-       ? (size_t)((GifFilePrivateType *)_gif->Private)->Read(_gif, _buf, _len) \
+#define READ(_gif, _buf, _len)                                         \
+  (((GifFilePrivateType *)_gif->Private)->Read                         \
+       ? ((GifFilePrivateType *)_gif->Private)->Read(_gif, _buf, _len) \
        : fread(_buf, 1, _len, ((GifFilePrivateType *)_gif->Private)->File))
 
 static int DGifGetWord(GifFileType *GifFile, GifWord *Word);
@@ -351,7 +351,7 @@ int DGifGetImageDesc(GifFileType *GifFile) {
     }
 
     /* Get the image local color map: */
-    for (i = 0; (int)i < GifFile->Image.ColorMap->ColorCount; i++) {
+    for (i = 0; i < (unsigned int)GifFile->Image.ColorMap->ColorCount; i++) {
       if (READ(GifFile, Buf, 3) != 3) {
         GifFreeMapObject(GifFile->Image.ColorMap);
         GifFile->Error = D_GIF_ERR_READ_FAILED;
